beest/testframework/stscliv1/requirements.txt
pytest-testinfra==6.6.0
marshmallow==3.17.0
